fix merge issues

+++ src/game_engine/board/display.rs
@@ -105,7 +105,8 @@ impl<B> Board for DisplayableBoard<B> where B: Board {
         self.inner.get_en_passant()
     }
 
-    fn get_material_score(&self) -> i8 {
+    #[inline]
+    fn get_material_score(&self) -> i32 {
         self.inner.get_material_score()
     }
